# Use Rails parallel tests for build improvements