Since this project uses minitest we can make use of the parallelisation feature provided by Rails. More details of the feature is available here: https://edgeguides.rubyonrails.org/testing.html#parallel-testing

I thought I may as well make the most of this being a minitest project before we do the RSpec rewrite - which hopefully will find lots of different ways to improve test performance (I've no idea what this suite does that takes so long since this project doesn't have a DB)

## Benchmarks

### On my local machine running govuk-docker

Before:

Finished in 800.161678s, 9.2006 runs/s, 25.1274 assertions/s.
7362 runs, 20106 assertions, 0 failures, 0 errors, 0 skips

After:

Finished in 216.857922s, 33.9485 runs/s, 92.7151 assertions/s.
7362 runs, 20106 assertions, 0 failures, 0 errors, 0 skips

### On CI

Before:

Finished in 659.463267s, 11.1636 runs/s, 30.4884 assertions/s.
7362 runs, 20106 assertions, 0 failures, 0 errors, 0 skips

After:

Finished in 94.375381s, 78.0076 runs/s, 213.0428 assertions/s.
7362 runs, 20106 assertions, 1 failures, 0 errors, 0 skips